+++ src/lib.c
@@ -3,12 +3,15 @@
 
 #include <Python.h>
 
+#define PyGEOS_API_Module
+
 #define PY_ARRAY_UNIQUE_SYMBOL pygeos_ARRAY_API
 #define PY_UFUNC_UNIQUE_SYMBOL pygeos_UFUNC_API
 #include <numpy/ndarraytypes.h>
 #include <numpy/npy_3kcompat.h>
 #include <numpy/ufuncobject.h>
 
+#include "c_api.h"
 #include "coords.h"
 #include "geos.h"
 #include "pygeom.h"
@@ -30,6 +33,9 @@ static struct PyModuleDef moduledef = {
 
 PyMODINIT_FUNC PyInit_lib(void) {
   PyObject *m, *d;
+
+  static void* PyGEOS_API[PyGEOS_API_num_pointers];
+  PyObject* c_api_object;
 
   m = PyModule_Create(&moduledef);
   if (!m) {
@@ -70,6 +76,16 @@ PyMODINIT_FUNC PyInit_lib(void) {
   if (init_ufuncs(m, d) < 0) {
     return NULL;
   };
+
+  /* Initialize the C API pointer array */
+  PyGEOS_API[PyGEOS_CreateGeometry_NUM] = (void*)PyGEOS_CreateGeometry;
+  PyGEOS_API[PyGEOS_GetGEOSGeometry_NUM] = (void*)PyGEOS_GetGEOSGeometry;
+
+  /* Create a Capsule containing the API pointer array's address */
+  c_api_object = PyCapsule_New((void*)PyGEOS_API, "pygeos.lib._C_API", NULL);
+  if (c_api_object != NULL) {
+    PyModule_AddObject(m, "_C_API", c_api_object);
+  }
 
   return m;
 }